+++ test/cases/_exceptions/msg-exception-scss-with-minify/webpack.config.js
@@ -1,0 +1,31 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+      css: {
+        filename: '[name].[contenthash:8].css',
+      },
+      minify: true, // test JS error if CSS has syntax error
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.s?css$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};